[REPEKA-100] Introduce static permissions to restrict access to system
Why commands are not authorized in the domain layer?
Well, it's not the domain responsibility to know whether the user
is allowed to perform the action or not. It is nicely described
here: http://stackoverflow.com/a/17161710/878514
Although I personally think that the authorization schema is
something strictly connected to the domain, it should not influence
the logic. Consider for example that the CreateUserCommand is
guarded by ADMIN role. How would you provide that role to the DI
when the command is sent from e.g. console?
Therefore, I agree that if user does not have the privilege to
execute some command/query he should not even be able to create it.

Change-Id: I78785c839d77e9147d655423163f6f38dbb0f211