[REPEKA-224] Fix nightly build
Including select2 in the vendor bundle triggers SystemJS's ES module
detection. SystemJS tries to transpile ES modules client-side, but we
have that feature disabled, so it gives up.

For some reason moving select2 into a separate bundle fixes this (sic!).

This fix should be considered a workaround until we move to Webpack.

Change-Id: Ia9e4d39cc3f75371f9243f2160006fb738a52839
Reviewed-on: https://gerrit.fslab.agh.edu.pl/7032
Tested-by: Jenkins Apps
Reviewed-by: Wojciech Frącz <fraczwojciech@gmail.com>