[REPEKA-281] Fix invalid behavior of repeat.for
Two versions of aurelia-binding were being installed, resulting in
massively buggy behavior.

The same problem was present with jQuery, which was installed both in
2.x and 3.x version. It didn't cause any problems so far because 2.x's
global $ was immediately overwritten by 3.x's $ before anything
interacted with it. After applying this fix some stuff started to happen
in different order and Bootstrap's plugins were loaded into jQuery 2.x
which was then overwritten by jQuery 3.x, which was unaware of any
plugins.

Another possibly problematic dependency was Cytoscape.js, which was also
installing in two versions (2.x and 3.x), but I haven't noticed any
issues with it. This one was preventively downgraded to 2.x because 3.x
was interacting with mouse clicks erroneously.

Change-Id: Ie07585d1c0d440d5c96a930f5992510e10d48af5