[REPEKA-301] Fix too generous bundle exclusions and disable @map by ref
The ability to provide @map with entity references is neat:

 @map(FooEntity) ...

But it also leads to issues after bundling, like class's static fields
being referenced before they are initialized (because the class is
declared later in the same bundle). Same with arrayOf() and dictOf()
helpers.

Usage of @map with explicit compile-time names is encouraged, without
any runtime stuff like FooEntity.NAME.

Change-Id: I13d5904d23587f197e9fba38463beebf2afb6b60