[REPEKA-355] Prevent overriding overrides of already overridden values
It turns out that EntityManager recycles entities, which is good in
most cases. Hovewer, with our custom listener that updates overrides
for each metadata in the resource kind it caused problems.

Image that in one request we fetch two resource kinds that uses the
same metadata with different overrides. Listener will call
updateOverrides() with overrides for each resource kind, but both
calls will refer to the same instance of metadata. This means that
the lastly set overrides will be the only ones that will be sent to
frontend.

Now, the updateOverrides methods has been renamed to withOverrides
and it always returns a cloned instance of Metadata to avoid
confusion.

The setOverrides() methods was deleted not to introduce confusion.

Change-Id: I4726078574ec2975dc5bd4f03d745d3cac432bfa