[REPEKA-651] Fix invalid constraints for reproductor
In production, they somehow managed to set overrides for reproductor.
These overrides set the resourceKind constraint to empty array,
therefore forbidding to choose any reproductor with "there is no
resource kind chosen for relationship" error message.

I did not manage to reproduce this problem, but the provided migration
fixes it by removing all of the overrides from the reproductor
metadata. The ResourceKindCreateCommandAdjuster now also ignores
all of the overrides that might have been given for the reproductor
metadata.

Change-Id: I12bc7019da5173d8e7978acfa355478b49f7c830