[REPEKA-708] Do not create objects when stripping metadata values
The previous REPEKA-708 commit introduced a very tricky bug.

1. User edits the worklow that has missing metadata ids e.g.
  in "requiredMetadataIds"; let's say it is
  "requiredMetadataIds" => [2, 6, 8]
2. Metadata "6" is missing, so the previous code filters it out.
3. The "requiredMetadataIds" now becomes [0 => 2, 2 => 8]
   but it should be [2, 8].

On the next workflow load, instead of
"requiredMetadataIds": [2, 8]
the GUI is being given
"requiredMetadataIds": {0: 2, 2: 8}
which it cannot understand and displays nothing for that place
in excel.

However, there is more than that. If user then edits such
workflow second time, backend receives no requiredMetadataIds.
In result, such action CLEARS ALL THE CONFIGURED METADATA IDS
from the places configuration.

This is sad, because this is exactly what happened in production.

:(

Change-Id: Iacbfaec4f3a8f86ab2b2143d8b1eace85e5c6898